test: split skill pack coverage tests
