+++ internal/service/skillpackservice/service_fakes_test.go
@@ -1,0 +1,306 @@
+package skillpackservice
+
+import (
+	"context"
+	"errors"
+	"io"
+	"net/http"
+	"strings"
+	"time"
+
+	"github.com/markhuangai/dense-mem/internal/domain"
+	"github.com/markhuangai/dense-mem/internal/http/dto"
+	"github.com/markhuangai/dense-mem/internal/service/claimservice"
+	"github.com/markhuangai/dense-mem/internal/service/factservice"
+	"github.com/markhuangai/dense-mem/internal/service/fragmentservice"
+	neo4jdriver "github.com/neo4j/neo4j-go-driver/v5/neo4j"
+)
+
+type fakeFactGet struct {
+	facts map[string]*domain.Fact
+}
+
+func packWithItem(sourceKind string) *SkillPack {
+	return &SkillPack{
+		SchemaVersion: SchemaVersion,
+		Name:          "Test pack",
+		Items: []SkillPackItem{{
+			Subject:    "assistant",
+			Predicate:  "has_skill",
+			Object:     "tests skill packs",
+			SourceKind: sourceKind,
+		}},
+	}
+}
+
+type fakeHTTPClient struct {
+	status     int
+	body       string
+	err        error
+	lastAccept string
+}
+
+func (f *fakeHTTPClient) Do(req *http.Request) (*http.Response, error) {
+	f.lastAccept = req.Header.Get("Accept")
+	if f.err != nil {
+		return nil, f.err
+	}
+	return &http.Response{
+		StatusCode: f.status,
+		Body:       io.NopCloser(strings.NewReader(f.body)),
+	}, nil
+}
+
+type mutatingScopedGraph struct {
+	reads int
+}
+
+func (g *mutatingScopedGraph) ScopedRead(_ context.Context, _ string, _ string, params map[string]any) (neo4jdriver.ResultSummary, []map[string]any, error) {
+	if _, exists := params["profileId"]; exists {
+		return nil, nil, errors.New("profileId parameter was reused")
+	}
+	params["profileId"] = "injected-by-scope"
+	g.reads++
+	if g.reads == 1 {
+		return nil, nil, nil
+	}
+	return nil, []map[string]any{{
+		"id":        "claim-1",
+		"type":      "claim",
+		"subject":   "assistant",
+		"predicate": "has_skill",
+		"object":    "skill pack regression testing",
+	}}, nil
+}
+
+func (g *mutatingScopedGraph) ScopedWrite(_ context.Context, _ string, _ string, _ map[string]any) (neo4jdriver.ResultSummary, error) {
+	return nil, nil
+}
+
+func (g *mutatingScopedGraph) ScopedWriteTx(_ context.Context, _ string, _ func(neo4jdriver.ManagedTransaction) error) error {
+	return nil
+}
+
+type factCandidateGraph struct {
+	reads int
+}
+
+func (g *factCandidateGraph) ScopedRead(_ context.Context, _ string, _ string, _ map[string]any) (neo4jdriver.ResultSummary, []map[string]any, error) {
+	g.reads++
+	return nil, []map[string]any{{
+		"id":        "fact-1",
+		"type":      "fact",
+		"subject":   "assistant",
+		"predicate": "has_skill",
+		"object":    "skill pack regression testing",
+	}}, nil
+}
+
+func (g *factCandidateGraph) ScopedWrite(_ context.Context, _ string, _ string, _ map[string]any) (neo4jdriver.ResultSummary, error) {
+	return nil, nil
+}
+
+func (g *factCandidateGraph) ScopedWriteTx(_ context.Context, _ string, _ func(neo4jdriver.ManagedTransaction) error) error {
+	return nil
+}
+
+type recordingGraph struct {
+	states     map[string]map[string]any
+	writeCount int
+	writeErr   error
+	txCount    int
+	readErr    error
+}
+
+func (g *recordingGraph) ScopedRead(_ context.Context, _ string, _ string, params map[string]any) (neo4jdriver.ResultSummary, []map[string]any, error) {
+	if g.readErr != nil {
+		return nil, nil, g.readErr
+	}
+	entityID, _ := params["entityId"].(string)
+	if entityID == "" {
+		return nil, nil, nil
+	}
+	state := g.states[entityID]
+	if state == nil {
+		return nil, nil, nil
+	}
+	return nil, []map[string]any{{"state": state}}, nil
+}
+
+func (g *recordingGraph) ScopedWrite(_ context.Context, _ string, _ string, _ map[string]any) (neo4jdriver.ResultSummary, error) {
+	g.writeCount++
+	return nil, g.writeErr
+}
+
+func (g *recordingGraph) ScopedWriteTx(_ context.Context, _ string, _ func(neo4jdriver.ManagedTransaction) error) error {
+	g.txCount++
+	return nil
+}
+
+func (f fakeFactGet) Get(_ context.Context, _ string, factID string) (*domain.Fact, error) {
+	return f.facts[factID], nil
+}
+
+type fakeClaimGet struct {
+	claims map[string]*domain.Claim
+}
+
+func (f fakeClaimGet) Get(_ context.Context, _ string, claimID string) (*domain.Claim, error) {
+	return f.claims[claimID], nil
+}
+
+type fakeFactList struct {
+	facts []*domain.Fact
+}
+
+func (f fakeFactList) List(_ context.Context, _ string, filters factservice.FactListFilters, _ int, _ string) ([]*domain.Fact, string, error) {
+	out := []*domain.Fact{}
+	for _, fact := range f.facts {
+		if filters.Subject != "" && fact.Subject != filters.Subject {
+			continue
+		}
+		if filters.Predicate != "" && fact.Predicate != filters.Predicate {
+			continue
+		}
+		if filters.Status != "" && fact.Status != filters.Status {
+			continue
+		}
+		out = append(out, fact)
+	}
+	return out, "", nil
+}
+
+type fakeClaimList struct {
+	claims []*domain.Claim
+}
+
+func (f fakeClaimList) List(_ context.Context, _ string, limit, offset int) ([]*domain.Claim, int, error) {
+	return f.claims, len(f.claims), nil
+}
+
+type fakeFragmentCreate struct {
+	calls int
+	err   error
+}
+
+func (f *fakeFragmentCreate) Create(_ context.Context, profileID string, req *dto.CreateFragmentRequest) (*fragmentservice.CreateResult, error) {
+	f.calls++
+	if f.err != nil {
+		return nil, f.err
+	}
+	return &fragmentservice.CreateResult{
+		Fragment: &domain.Fragment{
+			FragmentID:  "fragment-1",
+			ProfileID:   profileID,
+			Content:     req.Content,
+			ContentHash: "hash",
+			CreatedAt:   time.Now().UTC(),
+		},
+	}, nil
+}
+
+type fakeClaimCreate struct {
+	calls     int
+	duplicate bool
+	claimID   string
+	err       error
+}
+
+func (f *fakeClaimCreate) Create(_ context.Context, profileID string, claim *domain.Claim) (*claimservice.CreateResult, error) {
+	f.calls++
+	if f.err != nil {
+		return nil, f.err
+	}
+	claimID := f.claimID
+	if claimID == "" {
+		claimID = "claim-1"
+	}
+	claim.ClaimID = claimID
+	claim.ProfileID = profileID
+	claim.Status = domain.StatusCandidate
+	claim.EntailmentVerdict = domain.VerdictInsufficient
+	res := &claimservice.CreateResult{Claim: claim, Duplicate: f.duplicate}
+	if f.duplicate {
+		res.DuplicateOf = claimID
+	}
+	return res, nil
+}
+
+type fakeFactPromote struct {
+	err error
+}
+
+func (f fakeFactPromote) Promote(_ context.Context, profileID, claimID string) (*domain.Fact, error) {
+	if f.err != nil {
+		return nil, f.err
+	}
+	return &domain.Fact{
+		FactID:              "fact-1",
+		ProfileID:           profileID,
+		Subject:             "assistant",
+		Predicate:           "has_skill",
+		Object:              "tests skill packs",
+		Status:              domain.FactStatusActive,
+		PromotedFromClaimID: claimID,
+		RecordedAt:          time.Now().UTC(),
+	}, nil
+}
+
+type fakeLedger struct {
+	imports   []domain.SkillPackImport
+	changes   []domain.SkillPackImportChange
+	status    string
+	updateErr error
+	markErr   error
+	getErr    error
+	listErr   error
+	appendErr error
+}
+
+func (f *fakeLedger) CreateImport(_ context.Context, record domain.SkillPackImport) error {
+	f.imports = append(f.imports, record)
+	f.status = record.Status
+	return nil
+}
+
+func (f *fakeLedger) UpdateImportStatus(_ context.Context, _, _, status string, _, _ int, _ map[string]any) error {
+	if f.updateErr != nil {
+		return f.updateErr
+	}
+	f.status = status
+	return nil
+}
+
+func (f *fakeLedger) MarkRolledBack(context.Context, string, string) error {
+	if f.markErr != nil {
+		return f.markErr
+	}
+	f.status = domain.SkillPackImportStatusRolledBack
+	return nil
+}
+
+func (f *fakeLedger) GetImport(context.Context, string, string) (*domain.SkillPackImport, error) {
+	if f.getErr != nil {
+		return nil, f.getErr
+	}
+	status := f.status
+	if status == "" || status == domain.SkillPackImportStatusInspecting {
+		status = domain.SkillPackImportStatusApplied
+	}
+	return &domain.SkillPackImport{Status: status}, nil
+}
+
+func (f *fakeLedger) AppendChange(_ context.Context, change domain.SkillPackImportChange) error {
+	if f.appendErr != nil {
+		return f.appendErr
+	}
+	f.changes = append(f.changes, change)
+	return nil
+}
+
+func (f *fakeLedger) ListChanges(context.Context, string, string) ([]domain.SkillPackImportChange, error) {
+	if f.listErr != nil {
+		return nil, f.listErr
+	}
+	return f.changes, nil
+}
